feat: json export support
