+++ crates/iwe/Cargo.toml
@@ -34,6 +34,7 @@ tracing-tree.workspace = true
 graphviz-rust.workspace = true
 minijinja.workspace = true
 csv.workspace = true
+anyhow.workspace = true
 
 [dev-dependencies]
 tempfile.workspace = true
