+++ crates/iwe/src/export.rs
@@ -1,4 +1,5 @@
 pub mod dot_exporter;
-pub mod dot_details_exporter;
-pub mod graph_data;
+pub mod exporter;
 pub mod graph_colors;
+pub mod graph_data;
+pub mod json_exporter;
